+++ examples/lib/src/files_json.dart
@@ -1,4 +1,5 @@
 const List<String> filesJson = [
+  //'temp',
   'webgl_periodictable',
   'timing',
   'orientation',