+++ packages/three_js_core/lib/others/peripherals.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
@@ -34,6 +35,8 @@ class Peripherals extends StatefulWidget{
     return PeripheralsState();
   }
 }
+
+
 
 class PeripheralsState extends State<Peripherals> {
   final Map<PeripheralType, List<Function>> _listeners = {};
@@ -110,15 +113,27 @@ class PeripheralsState extends State<Peripherals> {
     _listeners[name] = cls;
   }
 
+  bool isSignal = false;
+  bool start = false;
+  Offset webPosition = Offset(0,0);
+
   @override
   Widget build(BuildContext context) {
     return KeyboardListener(
       focusNode: focusNode,
       onKeyEvent: (event){
         if(event is KeyDownEvent){
+          if(kIsWeb && start){
+            _onDragEvent(context, PeripheralType.pointerup, event);
+            start = false;
+          }
           _onKeyDownEvent(context, event.logicalKey);
         }
         else if(event is KeyUpEvent){
+          if(kIsWeb && start){
+            _onDragEvent(context, PeripheralType.pointerup, event);
+            start = false;
+          }
           _onKeyUpEvent(context, event.logicalKey);
         }
       },
@@ -135,11 +150,33 @@ class PeripheralsState extends State<Peripherals> {
         child: Listener(
           onPointerPanZoomStart: panGestureRecognizer.addPointerPanZoom,
           onPointerSignal: (event) {
-            if (event is PointerScrollEvent) {
-              _onPointerEvent(context, PeripheralType.wheel, event);
+            if(kIsWeb){
+              isSignal = true;
+              if (event is PointerScrollEvent) {
+                if(!start){
+                  webPosition = Offset(0, 0);
+                  _onDragEvent(context, PeripheralType.pointerdown, event);
+                  start = true;
+                }
+                else{
+                  webPosition-=event.scrollDelta;
+                  Map m = {
+                    'scrollDelta': webPosition,
+                    'position': event.position,
+                    'localPosition': event.localPosition
+                  };
+                  _onDragEvent(context, PeripheralType.pointermove, m);
+                }
+              }
+              else if(event is PointerScaleEvent){
+                double s = event.scale>1?-1:1;
+                _onScaleEvent(context, PeripheralType.wheel, {'scale':s});
+                _prevScale = event.scale;
+              }
             }
           },
           onPointerDown: (PointerDownEvent event) {
+            start = false;
             _pointers++;
             _onPointerEvent(context, PeripheralType.pointerdown, event);
             FocusScope.of(context).requestFocus(focusNode);
@@ -156,7 +193,16 @@ class PeripheralsState extends State<Peripherals> {
             _onPointerEvent(context, PeripheralType.pointercancel, event);
           },
           onPointerHover: (PointerHoverEvent event){
-            _onPointerEvent(context, PeripheralType.pointerHover , event);
+            if(!isSignal){
+              if(kIsWeb && start){
+                _onDragEvent(context, PeripheralType.pointerup, event);
+                start = false;
+              }
+              else{
+                _onPointerEvent(context, PeripheralType.pointerHover , event);
+              }
+            }
+            isSignal = false;
           },
           child: widget.builder(context),
         )
@@ -201,13 +247,13 @@ class WebPointerEvent {
   late int pointerId;
   late int button;
   String pointerType = 'touch';
-  late double clientX;
-  late double clientY;
-  late double pageX;
-  late double pageY;
+  double clientX = 0;
+  double clientY = 0;
+  double pageX = 0;
+  double pageY = 0;
 
-  late double movementX;
-  late double movementY;
+  double movementX = 0;
+  double movementY = 0;
 
   bool ctrlKey = false;
   bool metaKey = false;
@@ -247,7 +293,6 @@ class WebPointerEvent {
       final leftButtonPressed = event.buttons & 1 > 0;
       final rightButtonPressed = event.buttons & 2 > 0;
       final middleButtonPressed = event.buttons & 4 > 0;
-
       // Left button takes precedence over other
       if (leftButtonPressed) return 0;
       // 2nd priority is the right button
@@ -274,6 +319,7 @@ class WebPointerEvent {
     wpe.clientY = local.dy;
     wpe.pageX = event.position.dx;
     wpe.pageY = event.position.dy;
+    wpe.pointerCount = pointerCount;
 
     //if(event is PointerMoveEvent || event is PointerHoverEvent) {
       wpe.movementX = event.delta.dx;
@@ -307,36 +353,45 @@ class WebPointerEvent {
 
   static WebPointerEvent convertDragEvent(BuildContext context, event, int pointerCount) {
     final wpe = WebPointerEvent();
+    final EventTouch touch = EventTouch();
 
     wpe.pointerId = 512;
+    touch.pointer = 512;
+
     wpe.pointerType = 'touch_pad';
     wpe.button = 0;
     wpe.pointerCount = pointerCount;
+    
+    if(!kIsWeb){
+      final position = event.globalPosition;
+      RenderBox getBox = context.findRenderObject() as RenderBox;
+      final local = getBox.globalToLocal(position);
+      wpe.clientX = local.dx;
+      wpe.clientY = local.dy;
 
-    RenderBox getBox = context.findRenderObject() as RenderBox;
-    final local = getBox.globalToLocal(event.globalPosition);
-    wpe.clientX = local.dx;
-    wpe.clientY = local.dy;
+      wpe.pageX = position.dx;
+      wpe.pageY = position.dy;
 
-    wpe.pageX = event.globalPosition.dx;
-    wpe.pageY = event.globalPosition.dy;
-    if(event is! DragStartDetails && event is! DragEndDetails){
-      wpe.movementX = event.delta.dx;
-      wpe.movementY = event.delta.dy;
+      wpe.deltaX = local.dx;
+      wpe.deltaY = local.dy;
+
+      touch.pageX = event.localPosition.dx;
+      touch.pageY = event.localPosition.dy;
+
+      touch.clientX = event.localPosition.dx;
+      touch.clientY = event.localPosition.dy;
     }
+    else if(event is Map){
+      final local = event['scrollDelta'];
+      wpe.clientX = local.dx;
+      wpe.clientY = local.dy;
 
-    wpe.deltaX = event.globalPosition.dx;
-    wpe.deltaY = event.globalPosition.dy;
-    wpe.clientX = event.globalPosition.dx;
-    wpe.clientY = event.globalPosition.dy;
+      wpe.pageX = event['position'].dx;
+      wpe.pageY = event['position'].dy;
 
-    final EventTouch touch = EventTouch();
-
-    touch.pointer = 512;
-    touch.pageX = event.localPosition.dx;
-    touch.pageY = event.localPosition.dy;
-    touch.clientX = event.localPosition.dx;
-    touch.clientY = event.localPosition.dy;
+      wpe.deltaX = local.dx;
+      wpe.deltaY = local.dy;
+    }
 
     wpe.touches.add(touch);
     wpe.changedTouches = [touch];